ImportC: unused `__forceinline` and `inline` functions still generate code
Introduce a notion of `cInline` for a function so that we can elide
emitting it if it is never referenced. Note that this is orthogonal
to `pragma(inline)` semantics, which is an optimization hint.

A cInline function will be output in the object file of the first
module that uses it, not the module that defines it. If it is
never used, then it won't be emitted.

Avoiding emitting these inline functions is necessary for the common
practice in C of referencing potentially unresolved external symbols
from an inline function.